+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1AJRSG8FMFM6J2DYTZ81FT"
+        "01KN1EM1GV7Q9Y0VDKXJTEWF3S"
     )
     r"""The unique identifier of the data source"""