+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1AJRSJPCM6QY0MQG6WNS58"
+        "01KN1EM1GXR6HBB1XC83SJXZ7F"
     )
     r"""The unique identifier of the data source"""